% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_baldwin_mutations.R
\name{apply_baldwin_mutations}
\alias{apply_baldwin_mutations}
\title{apply_baldwin_mutations}
\usage{
apply_baldwin_mutations(p)
}
\arguments{
\item{p}{input dataframe}
}
\description{
apply_baldwin_mutations
}
