% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_colors_and_logos.R
\name{apply_colors_and_logos}
\alias{apply_colors_and_logos}
\title{apply_colors_and_logos}
\usage{
apply_colors_and_logos(p, team_col = "")
}
\arguments{
\item{p}{an input dataframe}

\item{team_col}{a column specifying team name/abbreviation}
}
\description{
apply_colors_and_logos
}
