% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_completion_probability.R
\name{apply_completion_probability}
\alias{apply_completion_probability}
\title{Title}
\usage{
apply_completion_probability(p, all_plays)
}
\arguments{
\item{p}{input dataframe}

\item{all_plays}{all play by play data since 2009}
}
\description{
Title
}
