% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_sharpe_mutations.R
\name{apply_sharpe_mutations}
\alias{apply_sharpe_mutations}
\title{apply_sharp_mutations}
\usage{
apply_sharpe_mutations(p)
}
\arguments{
\item{p}{Input dataframe}
}
\description{
apply_sharp_mutations
}
