% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_games.R
\name{double_games}
\alias{double_games}
\title{double_games}
\usage{
double_games(g)
}
\arguments{
\item{g}{input dataframe}
}
\description{
takes input where each game has one row with teams as `away_team` and `home_team`
returns with each game having two rows with teams listed as `team` and `opp`
}
