% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_inconsistent_data_types.R
\name{fix_inconsistent_data_types}
\alias{fix_inconsistent_data_types}
\title{fix_inconsistent_data_types}
\usage{
fix_inconsistent_data_types(p)
}
\arguments{
\item{p}{Input dataframe}
}
\description{
fix_inconsistent_data_types
}
