% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_team_abbreviations.R
\name{fix_team_abbreviations}
\alias{fix_team_abbreviations}
\title{fix_team_abbreviations}
\usage{
fix_team_abbreviations(p, old_to_new = FALSE)
}
\arguments{
\item{p}{Input dataframe}

\item{old_to_new}{convert old name to new name (TRUE/FALSE)}
}
\description{
fix_team_abbreviations
}
