% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{grep_col}
\alias{grep_col}
\title{grep_col
look for text in names of columns of data frame}
\usage{
grep_col(x, df = plays)
}
\arguments{
\item{df}{}
}
\description{
grep_col
look for text in names of columns of data frame
}
